+++ upgrade/upgrade-1.1.0.php
@@ -1,0 +1,10 @@
+<?php
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+function upgrade_module_1_1_0($module)
+{
+    return $module instanceof Wilden_manager;
+}